fix(files_sharing): Apply default password setting in SharingDetailsTab
