+++ apps/files_sharing/src/mixins/SharesMixin.js
@@ -11,6 +11,7 @@ import { emit } from '@nextcloud/event-bus'
 import PQueue from 'p-queue'
 import debounce from 'debounce'
 
+import GeneratePassword from '../utils/GeneratePassword.ts'
 import Share from '../models/Share.ts'
 import SharesRequests from './ShareRequests.js'
 import Config from '../services/ConfigService.ts'
@@ -155,6 +156,26 @@ export default {
 				return this.config.defaultInternalExpirationDate
 			}
 			return null
+		},
+		/**
+		 * Is the current share password protected ?
+		 *
+		 * @return {boolean}
+		 */
+		isPasswordProtected: {
+			get() {
+				return this.config.enforcePasswordForPublicLink
+							|| !!this.share.password
+			},
+			async set(enabled) {
+				if (enabled) {
+					this.share.password = await GeneratePassword(true)
+					this.$set(this.share, 'newPassword', this.share.password)
+				} else {
+					this.share.password = ''
+					this.$delete(this.share, 'newPassword')
+				}
+			},
 		},
 	},
